relocate torch optimization and fix seed.
