+++ language/gpt-j/utils.py
@@ -1,6 +1,35 @@
 import json
 import os
 import io
+import torch
+import numpy as np
+import random 
+
+def random_seed(seed=42):
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)
+    np.random.seed(seed)
+    random.seed(seed)
+
+def set_optimization(args):
+    if args.torch_optim == 'default':
+        return
+    elif args.torch_optim == 'none':
+        torch.backends.cuda.matmul.allow_tf32 = False
+        torch.backends.cuda.matmul.allow_fp16_reduced_precision_reduction = False
+        torch.backends.cuda.matmul.allow_bf16_reduced_precision_reduction = False
+        torch.backends.cuda.enable_mem_efficient_sdp(False)
+        torch.backends.cuda.enable_flash_sdp(False)
+        torch.backends.cudnn.enabled = False
+        torch.backends.cudnn.allow_tf32 = False
+        torch.backends.cudnn.deterministic = True
+        torch.backends.cudnn.benchmark = False
+        if hasattr(torch.backends, 'opt_einsum'):
+            torch.backends.opt_einsum.enabled = False
+    else:
+        raise ValueError(f"Wrong argument value for '--torch_optim': {args.torch_optim}")
+
+    return
 
 def _make_r_io_base(f, mode: str):
     if not isinstance(f, io.IOBase):
